problem solution added of Lucky Numbers

+++ Module 6.5/D_Ali_Baba_and_Puzzles.c
@@ -2,7 +2,7 @@
 
 #include<stdio.h>
 int main(){
-    long long int a,b,c;
+    int a,b,c;
     long long int d;
     scanf("%lld %lld %lld %lld",&a, &b, &c, &d);
     if (a + b - c == d || a - b + c == d ||  a + (b * c) == d || (a * b) + c  == d || a - (b * c) == d || (a * b) - c == d)
